restructing the codebase
