+++ utils/constants/constants.go
@@ -1,0 +1,11 @@
+package constants
+
+var (
+	MapValues = map[string]string{
+		"Sun Pharmaceutical Industries Limited":       "Sun Pharma.Inds.",
+		"KEC International Limited":                   "K E C Intl.",
+		"Sandhar Technologies Limited":                "Sandhar Tech",
+		"Samvardhana Motherson International Limited": "Samvardh. Mothe.",
+		"Coromandel International Limited":            "Coromandel Inter",
+	}
+)
